Pull out EC2 signature unwrapping into helper
